fix: improve survey bundle size (#1759)
