+++ src/posthog-surveys.ts
@@ -1,13 +1,7 @@
 import { SURVEYS } from './constants'
 import { getSurveySeenStorageKeys } from './extensions/surveys/surveys-utils'
 import { PostHog } from './posthog-core'
-import {
-    Survey,
-    SurveyCallback,
-    SurveyMatchType,
-    SurveyQuestionBranchingType,
-    SurveyQuestionType,
-} from './posthog-surveys-types'
+import { Survey, SurveyCallback, SurveyMatchType } from './posthog-surveys-types'
 import { RemoteConfig } from './types'
 import { Info } from './utils/event-utils'
 import { assignableWindow, document, userAgent, window } from './utils/globals'
@@ -30,109 +24,6 @@ export const surveyValidationMap: Record<SurveyMatchType, (targets: string[], va
     exact: (targets, value) => targets.some((target) => value === target),
 
     is_not: (targets, value) => targets.every((target) => value !== target),
-}
-
-function getRatingBucketForResponseValue(responseValue: number, scale: number) {
-    if (scale === 3) {
-        if (responseValue < 1 || responseValue > 3) {
-            throw new Error('The response must be in range 1-3')
-        }
-
-        return responseValue === 1 ? 'negative' : responseValue === 2 ? 'neutral' : 'positive'
-    } else if (scale === 5) {
-        if (responseValue < 1 || responseValue > 5) {
-            throw new Error('The response must be in range 1-5')
-        }
-
-        return responseValue <= 2 ? 'negative' : responseValue === 3 ? 'neutral' : 'positive'
-    } else if (scale === 7) {
-        if (responseValue < 1 || responseValue > 7) {
-            throw new Error('The response must be in range 1-7')
-        }
-
-        return responseValue <= 3 ? 'negative' : responseValue === 4 ? 'neutral' : 'positive'
-    } else if (scale === 10) {
-        if (responseValue < 0 || responseValue > 10) {
-            throw new Error('The response must be in range 0-10')
-        }
-
-        return responseValue <= 6 ? 'detractors' : responseValue <= 8 ? 'passives' : 'promoters'
-    }
-
-    throw new Error('The scale must be one of: 3, 5, 7, 10')
-}
-
-export function getNextSurveyStep(
-    survey: Survey,
-    currentQuestionIndex: number,
-    response: string | string[] | number | null
-) {
-    const question = survey.questions[currentQuestionIndex]
-    const nextQuestionIndex = currentQuestionIndex + 1
-
-    if (!question.branching?.type) {
-        if (currentQuestionIndex === survey.questions.length - 1) {
-            return SurveyQuestionBranchingType.End
-        }
-
-        return nextQuestionIndex
-    }
-
-    if (question.branching.type === SurveyQuestionBranchingType.End) {
-        return SurveyQuestionBranchingType.End
-    } else if (question.branching.type === SurveyQuestionBranchingType.SpecificQuestion) {
-        if (Number.isInteger(question.branching.index)) {
-            return question.branching.index
-        }
-    } else if (question.branching.type === SurveyQuestionBranchingType.ResponseBased) {
-        // Single choice
-        if (question.type === SurveyQuestionType.SingleChoice) {
-            // :KLUDGE: for now, look up the choiceIndex based on the response
-            // TODO: once QuestionTypes.MultipleChoiceQuestion is refactored, pass the selected choiceIndex into this method
-            const selectedChoiceIndex = question.choices.indexOf(`${response}`)
-
-            if (question.branching?.responseValues?.hasOwnProperty(selectedChoiceIndex)) {
-                const nextStep = question.branching.responseValues[selectedChoiceIndex]
-
-                // Specific question
-                if (Number.isInteger(nextStep)) {
-                    return nextStep
-                }
-
-                if (nextStep === SurveyQuestionBranchingType.End) {
-                    return SurveyQuestionBranchingType.End
-                }
-
-                return nextQuestionIndex
-            }
-        } else if (question.type === SurveyQuestionType.Rating) {
-            if (typeof response !== 'number' || !Number.isInteger(response)) {
-                throw new Error('The response type must be an integer')
-            }
-
-            const ratingBucket = getRatingBucketForResponseValue(response, question.scale)
-
-            if (question.branching?.responseValues?.hasOwnProperty(ratingBucket)) {
-                const nextStep = question.branching.responseValues[ratingBucket]
-
-                // Specific question
-                if (Number.isInteger(nextStep)) {
-                    return nextStep
-                }
-
-                if (nextStep === SurveyQuestionBranchingType.End) {
-                    return SurveyQuestionBranchingType.End
-                }
-
-                return nextQuestionIndex
-            }
-        }
-
-        return nextQuestionIndex
-    }
-
-    logger.warn('Falling back to next question index due to unexpected branching type')
-    return nextQuestionIndex
 }
 
 function defaultMatchType(matchType?: SurveyMatchType): SurveyMatchType {
@@ -402,7 +293,6 @@ export class PostHogSurveys {
             return this.instance.featureFlags.isFeatureEnabled(value)
         })
     }
-    getNextSurveyStep = getNextSurveyStep
 
     // this method is lazily loaded onto the window to avoid loading preact and other dependencies if surveys is not enabled
     private _canActivateRepeatedly(survey: Survey) {
